pyproject.toml
[tool.poetry]
name = "segmentation-task"
version = "0.1.0"
description = ""
authors = ["SotaChambers <snakamura@ridge-i.com>"]
readme = "README.md"
packages = [{include = "segmentation_task"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
ipykernel = "^6.19.4"
segmentation-models-pytorch = "^0.3.1"
matplotlib = "^3.6.2"
albumentations = "^1.3.0"
flake8 = "^6.0.0"
pandas = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
